Add files via upload
## Contribution Guidelines

This file outlines the process and expectations for contributing to the IHCRF GitHub repository. Whether you're submitting code, reporting an issue, or providing documentation, these guidelines provide the steps to ensure your contributions are clear, effective, and aligned with the goals of the India Health and Climate Resilience Fellowship (IHCRF).

Refer to this file to understand how to:
- Report issues or bugs.
- Submit new code or improvements.
- Contribute to documentation or suggest new features.
- Follow our coding standards and best practices.

We welcome contributions from the global community to help enhance our work in advancing health and climate resilience.